SPLICE-927: Use Batch Nested Loop Join Operation for all NestedLoop Joins
