+++ splice_machine/src/main/java/com/splicemachine/derby/stream/function/NLJOuterJoinFunction.java
@@ -15,28 +15,21 @@
 
 package com.splicemachine.derby.stream.function;
 
-import com.splicemachine.EngineDriver;
-import com.splicemachine.db.iapi.sql.execute.ExecRow;
+import com.splicemachine.db.iapi.error.StandardException;
 import com.splicemachine.derby.iapi.sql.execute.SpliceOperation;
-import com.splicemachine.derby.impl.sql.execute.operations.JoinUtils;
 import com.splicemachine.derby.impl.sql.execute.operations.LocatedRow;
-import com.splicemachine.derby.stream.iapi.DataSet;
 import com.splicemachine.derby.stream.iapi.OperationContext;
-import com.splicemachine.derby.stream.iterator.NestedLoopLeftOuterIterator;
-import com.splicemachine.derby.stream.utils.StreamLogUtils;
-import com.splicemachine.derby.stream.utils.StreamUtils;
+import com.splicemachine.derby.stream.iterator.NestedLoopJoinIterator;
+
 import java.io.IOException;
 import java.io.ObjectInput;
 import java.io.ObjectOutput;
-import java.util.Collections;
 import java.util.Iterator;
 
 /**
  * Created by jleach on 4/24/15.
  */
-public class NLJOuterJoinFunction<Op extends SpliceOperation> extends SpliceJoinFlatMapFunction<Op, LocatedRow, LocatedRow> {
-    public Iterator<LocatedRow> rightSideNLJIterator;
-    public LocatedRow leftRow;
+public class NLJOuterJoinFunction<Op extends SpliceOperation> extends NLJoinFunction<Op, Iterator<LocatedRow>, LocatedRow> {
 
     public NLJOuterJoinFunction() {}
 
@@ -55,27 +48,17 @@ public class NLJOuterJoinFunction<Op extends SpliceOperation> extends SpliceJoin
     }
 
     @Override
-    public Iterable<LocatedRow> call(LocatedRow from) throws Exception {
-        checkInit();
-        DataSet dataSet = null;
-        try {
-            leftRow = from;
-            op.getRightOperation().openCore(EngineDriver.driver().processorFactory().localProcessor(null,op));
-            rightSideNLJIterator = op.getRightOperation().getLocatedRowIterator();
-            if (!rightSideNLJIterator.hasNext()) {
-                // No Rows Right Side...
-                ExecRow mergedRow = JoinUtils.getMergedRow(leftRow.getRow(),
-                        op.getEmptyRow(), op.wasRightOuterJoin
-                        , executionFactory.getValueRow(numberOfColumns));
-                LocatedRow lr = new LocatedRow(leftRow.getRowLocation(),mergedRow.getClone());
-                StreamLogUtils.logOperationRecordWithMessage(lr,operationContext,"outer - right side no rows");
-                op.setCurrentLocatedRow(lr);
-                return Collections.singletonList(lr);
-            }
-            return new NestedLoopLeftOuterIterator(this);
-        } finally {
-            if (op.getRightOperation()!= null)
-                op.getRightOperation().close();
+    public Iterable<LocatedRow> call(Iterator<LocatedRow> from) throws Exception {
+        if (!initialized) {
+            init(from);
+            initialized = true;
         }
+
+        return new NestedLoopJoinIterator<>(this);
+    }
+
+    protected void init(Iterator<LocatedRow> from) throws StandardException {
+        joinType = JoinType.LEFT_OUTER;
+        super.init(from);
     }
 }
